Unifying data by introducing serialization method
This commit introduces a serialize method for all database objects
making it easier to get a unified representation of the same data.

Additionally some parts of the database were restructured. This hold
especially for the relationship between player and playlist with the
player object now holding the song currently playing.

Signed-off-by: Lars Kiesow <lkiesow@uos.de>